Move more default classes to correct namespace
